Make database.interval_depth uint16_t (vs 8) (boost bug).

+++ src/parser.cpp
@@ -1335,7 +1335,7 @@ options_metadata parser::load_settings() THROWS
     )
     (
         "database.interval_depth",
-        value<uint8_t>(&configured.database.interval_depth),
+        value<uint16_t>(&configured.database.interval_depth),
         "The interval depth for merkle proof optimization, defaults to '255' (disabled)."
     )
